+++ test/Main.hs
@@ -25,6 +25,7 @@ import Effectful.Tracing.NoOpSpec qualified as NoOpSpec
 import Effectful.Tracing.PrettyPrintLeakSpec qualified as PrettyPrintLeakSpec
 import Effectful.Tracing.PrettyPrintSpec qualified as PrettyPrintSpec
 import Effectful.Tracing.Propagation.B3Spec qualified as B3Spec
+import Effectful.Tracing.Propagation.JaegerSpec qualified as JaegerSpec
 import Effectful.Tracing.PropagationSpec qualified as PropagationSpec
 import Effectful.Tracing.PropertySpec qualified as PropertySpec
 import Effectful.Tracing.SamplerSpec qualified as SamplerSpec
@@ -65,6 +66,7 @@ main =
           , ConcurrentSpec.tests
           , PropagationSpec.tests
           , B3Spec.tests
+          , JaegerSpec.tests
           , BaggageSpec.tests
           , FuzzSpec.tests
           , ThunkSpec.tests